Update AmdCbs file that fixes xgbe and sata enumeration
It seems that Insyde's 05.38.09.0023 bios which provides Linux support
had some functional differences in the AmdCbsForm.vfr file.

Arista's changes in the AmdCbsCustomization folder for the AristaPuffin
project was not setting the XGBE0 and SATA controller options to DEFAULT
in the AmdCbsForm.vfr file.

This seemed to not matter in the older version of this bios but it seems
once we migrated the project to the new bios the issue started taking
place.

With this change the sata and xgbe devices are enumerated on pcie.

Signed-off-by: Gianfranco <gianfranco.dutka@arista.com>